remove: exam, studyLog 패키지를 domain 이하로 이동
