+++ src/main/java/com/example/masterplanbbe/common/exception/ErrorCode.java
@@ -29,7 +29,7 @@ public enum ErrorCode {
     // Post
     INVALID_INPUT_TITLE(400,"P001","유효하지 않은 타이틀입니다."),
     INVALID_INPUT_CONTENT(400,"P002","유효하지 않은 내용입니다."),
-    NOT_FOUND_POST(400,"P003" ,"게시글을 찾을 수 없습니다." ),
+    NOT_FIND_POST(400,"P003" ,"게시글을 찾을 수 없습니다." ),
     NOT_DELETED_POST(400,"P004","게시글 삭제 권한이 없습니다."),
     NOT_MODIFIED_POST(400,"P005" ,"게시글 수정 권한이 없습니다." ),
