+++ src/main/java/com/example/masterplanbbe/domain/exam/enums/Category.java
@@ -1,0 +1,14 @@
+package com.example.masterplanbbe.domain.exam.enums;
+
+//TODO: 상호배타적이지 않아야 한다. 비트 플래그 방식이 후보
+public enum Category {
+    IT_ICT,
+    FINANCE_INSURANCE,
+    BUSINESS_ACCOUNTING,
+    HEALTH_MEDICAL,
+    EDUCATION_SOCIAL_SCIENCE,
+    ARTS_DESIGN,
+    CONSTRUCTION_ENGINEERING,
+    ELECTRICAL_ELECTRONICS,
+    LANGUAGE
+}
