+++ src/main/java/com/example/masterplanbbe/domain/exam/enums/CertificationType.java
@@ -1,0 +1,9 @@
+package com.example.masterplanbbe.domain.exam.enums;
+
+public enum CertificationType {
+    NATIONAL_TECHNICAL,
+    NATIONAL_PROFESSIONAL,
+    NATIONAL_CERTIFIED,
+    REGISTERED_PRIVATE,
+    SKILL_ASSESSMENT
+}
