+++ src/main/java/com/example/masterplanbbe/domain/exam/enums/Difficulty.java
@@ -1,0 +1,5 @@
+package com.example.masterplanbbe.domain.exam.enums;
+
+public enum Difficulty {
+
+}
