+++ src/main/java/com/example/masterplanbbe/domain/exam/enums/SessionType.java
@@ -1,0 +1,7 @@
+package com.example.masterplanbbe.domain.exam.enums;
+
+public enum SessionType {
+    WRITTEN,
+    PRACTICAL,
+    NONE
+}
