+++ src/main/java/com/example/masterplanbbe/domain/exam/repository/ExamBookmarkRepository.java
@@ -1,0 +1,7 @@
+package com.example.masterplanbbe.domain.exam.repository;
+
+import com.example.masterplanbbe.domain.exam.entity.ExamBookmark;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface ExamBookmarkRepository extends JpaRepository<ExamBookmark, Long> {
+}
