+++ src/main/java/com/example/masterplanbbe/domain/exam/repository/ExamRepository.java
@@ -1,0 +1,7 @@
+package com.example.masterplanbbe.domain.exam.repository;
+
+import com.example.masterplanbbe.domain.exam.entity.Exam;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface ExamRepository extends JpaRepository<Exam, Long>, ExamRepositoryCustom {
+}
