+++ src/main/java/com/example/masterplanbbe/domain/exam/repository/ExamRepositoryCustom.java
@@ -1,0 +1,9 @@
+package com.example.masterplanbbe.domain.exam.repository;
+
+import com.example.masterplanbbe.domain.exam.dto.ExamItemCardDto;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
+
+public interface ExamRepositoryCustom {
+    Page<ExamItemCardDto> getExamItemCards(Pageable pageable, String memberId);
+}
