+++ src/main/java/com/example/masterplanbbe/domain/exam/response/ReadAllExamResponse.java
@@ -1,0 +1,8 @@
+package com.example.masterplanbbe.domain.exam.response;
+
+import com.example.masterplanbbe.domain.exam.dto.ExamItemCardDto;
+import org.springframework.data.domain.Page;
+
+public record ReadAllExamResponse(
+        Page<ExamItemCardDto> examItemCardDtoPage
+) {}
