+++ src/main/java/com/example/masterplanbbe/domain/exam/response/ReadExamResponse.java
@@ -1,0 +1,4 @@
+package com.example.masterplanbbe.domain.exam.response;
+
+public record ReadExamResponse() {
+}
