+++ src/main/java/com/example/masterplanbbe/domain/studyLog/enums/InputSource.java
@@ -1,0 +1,6 @@
+package com.example.masterplanbbe.domain.studyLog.enums;
+
+public enum InputSource {
+    COMPUTER,
+    MOBILE
+}
