+++ src/main/java/com/example/masterplanbbe/domain/studyLog/repository/StudyLogRepository.java
@@ -1,0 +1,7 @@
+package com.example.masterplanbbe.domain.studyLog.repository;
+
+import com.example.masterplanbbe.domain.studyLog.entity.StudyLog;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface StudyLogRepository extends JpaRepository<StudyLog, Long>, StudyLogRepositoryCustom {
+}
