+++ src/main/java/com/example/masterplanbbe/domain/studyLog/repository/StudyLogRepositoryCustom.java
@@ -1,0 +1,5 @@
+package com.example.masterplanbbe.domain.studyLog.repository;
+
+public interface StudyLogRepositoryCustom {
+    // JpaRepository 가 생성하지 않는, 커스텀 메소드를 정의하는 인터페이스
+}
