+++ src/main/java/com/example/masterplanbbe/domain/studyLog/repository/StudyLogRepositoryImpl.java
@@ -1,0 +1,4 @@
+package com.example.masterplanbbe.domain.studyLog.repository;
+
+public class StudyLogRepositoryImpl implements StudyLogRepositoryCustom {
+}
