--- src/main/java/com/example/masterplanbbe/exam/entity/ExamBookmark.java
@@ -1,24 +1,0 @@
-package com.example.masterplanbbe.exam.entity;
-
-import com.example.masterplanbbe.common.annotation.NonNull;
-import com.example.masterplanbbe.common.domain.IdAndCreatedEntity;
-import com.example.masterplanbbe.member.entity.Member;
-import jakarta.persistence.*;
-import lombok.*;
-
-@Entity
-@Table(name = "exam_bookmarks")
-@Getter
-@AllArgsConstructor
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class ExamBookmark extends IdAndCreatedEntity {
-    @NonNull
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "member_id")
-    private Member member;
-
-    @NonNull
-    @OneToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "exam_id")
-    private Exam exam;
-}
