--- src/main/java/com/example/masterplanbbe/exam/enums/CertificationType.java
@@ -1,9 +1,0 @@
-package com.example.masterplanbbe.exam.enums;
-
-public enum CertificationType {
-    NATIONAL_TECHNICAL,
-    NATIONAL_PROFESSIONAL,
-    NATIONAL_CERTIFIED,
-    REGISTERED_PRIVATE,
-    SKILL_ASSESSMENT
-}
