--- src/main/java/com/example/masterplanbbe/exam/enums/Difficulty.java
@@ -1,5 +1,0 @@
-package com.example.masterplanbbe.exam.enums;
-
-public enum Difficulty {
-
-}
