--- src/main/java/com/example/masterplanbbe/exam/enums/SessionType.java
@@ -1,7 +1,0 @@
-package com.example.masterplanbbe.exam.enums;
-
-public enum SessionType {
-    WRITTEN,
-    PRACTICAL,
-    NONE
-}
