--- src/main/java/com/example/masterplanbbe/exam/repository/ExamBookmarkRepository.java
@@ -1,7 +1,0 @@
-package com.example.masterplanbbe.exam.repository;
-
-import com.example.masterplanbbe.exam.entity.ExamBookmark;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface ExamBookmarkRepository extends JpaRepository<ExamBookmark, Long> {
-}
