--- src/main/java/com/example/masterplanbbe/exam/repository/ExamRepository.java
@@ -1,7 +1,0 @@
-package com.example.masterplanbbe.exam.repository;
-
-import com.example.masterplanbbe.exam.entity.Exam;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface ExamRepository extends JpaRepository<Exam, Long>, ExamRepositoryCustom {
-}
