--- src/main/java/com/example/masterplanbbe/exam/response/ReadAllExamResponse.java
@@ -1,8 +1,0 @@
-package com.example.masterplanbbe.exam.response;
-
-import com.example.masterplanbbe.exam.dto.ExamItemCardDto;
-import org.springframework.data.domain.Page;
-
-public record ReadAllExamResponse(
-        Page<ExamItemCardDto> examItemCardDtoPage
-) {}
