--- src/main/java/com/example/masterplanbbe/exam/response/ReadExamResponse.java
@@ -1,4 +1,0 @@
-package com.example.masterplanbbe.exam.response;
-
-public record ReadExamResponse() {
-}
