--- src/main/java/com/example/masterplanbbe/studyLog/enums/InputSource.java
@@ -1,6 +1,0 @@
-package com.example.masterplanbbe.studyLog.enums;
-
-public enum InputSource {
-    COMPUTER,
-    MOBILE
-}
