--- src/main/java/com/example/masterplanbbe/studyLog/repository/StudyLogRepositoryCustom.java
@@ -1,5 +1,0 @@
-package com.example.masterplanbbe.studyLog.repository;
-
-public interface StudyLogRepositoryCustom {
-    // JpaRepository 가 생성하지 않는, 커스텀 메소드를 정의하는 인터페이스
-}
