--- src/main/java/com/example/masterplanbbe/studyLog/repository/StudyLogRepositoryImpl.java
@@ -1,4 +1,0 @@
-package com.example.masterplanbbe.studyLog.repository;
-
-public class StudyLogRepositoryImpl implements StudyLogRepositoryCustom {
-}
